Fix: Refactor API structure and views arch pattern
