+++ src/core/models/answers.ts
@@ -1,11 +1,11 @@
-import { QuestionnaireResult } from "../../shared/types/questionnaire";
+import { IQuestionnaireResult } from "../../shared/types/questionnaire";
 
-export interface WebAnswersDTO {
+export interface IWebAnswer {
     userId: number;
-    questionnaireData: QuestionnaireResult;
+    questionnaireData: IQuestionnaireResult;
 }
 
-export interface AIRecommendationResult {
+export interface IRecommendationAI {
     resumenUsuario: string;
     colorSemaforo: string; // 'verde' | 'amarillo' | 'rojo'
     analisisAsesor?: string;
