+++ src/pages/admin/Reports.tsx
@@ -3,8 +3,10 @@ import { IAnalysis } from "../../core/models/analysis";
 import { IQuestionnaire } from "../../core/models/questionnaire";
 import { IUser } from "../../core/models/user";
 import { IQuestion } from "../../core/models/question";
-import { fetchQuestionsByQuestionnaire, getAllAnalysis, getAllQuestionnaires } from "../../api/analysisApi";
-import { getAllUsers } from "../../api/userApi";
+import { getAllQuestionnaires } from "../../api/questionnairesApi";
+import { getAllAnalysis } from "../../api/analysisApi";
+import { fetchQuestionsByQuestionnaire } from "../../api/questionnairesApi";
+import { getAllUsers } from "../../api/usersApi";
 import { X, Search, ClipboardList, FileText, Users } from "lucide-react";
 import TemplateMetrics from "../../shared/ui/template/TemplateMetrics";
