+++ src/pages/adviser/Profile.tsx
@@ -5,11 +5,11 @@ import {
 } from 'lucide-react';
 import { useAuth } from '../../shared/context/AuthContext';
 import { jwtDecode } from 'jwt-decode';
-import { getUserById, deleteUserById } from '../../api/userApi';
+import { getUserById, deleteUserById } from '../../api/usersApi';
 import { useNavigate } from 'react-router-dom';
 import { IUser } from '../../core/models/user';
 import ModalEditUser from '../../shared/ui/components/modals/ModalEditUser';
-import CardConfirmDelete from '../../shared/ui/components/cards/CardConfirmDelete';
+import DialogConfirmDelete from '../../shared/ui/components/dialogs/DialogConfirmDelete';
 
 interface JwtPayload {
   sub: string; // email
@@ -207,7 +207,7 @@ function AdviserProfile() {
 
       {/* Modal de confirmación de eliminación */}
       {showDeleteConfirm && (
-        <CardConfirmDelete
+        <DialogConfirmDelete
           title="Eliminar cuenta"
           message="¿Estás seguro de que deseas eliminar tu cuenta? Esta acción no se puede deshacer."
           confirmText="Sí, eliminar"
