+++ src/shared/types/questionnaire.ts
@@ -1,24 +1,24 @@
 import { IQuestion, QuestionType } from "../../core/models/question";
 
-export interface QuestionnaireData {
+export interface IQuestionnaireData {
   id: string;
   questions: IQuestion[];
 }
 
-export interface QuestionnaireMetadata {
+export interface IQuestionnaireMetadata {
   category: string;
   clientType: string;
   timestamp: string;
 }
 
-export interface QuestionnaireAnswer {
+export interface IQuestionnaireAnswer {
   questionId: number;
   questionTitle: string;
   answer: string[];
   questionType: QuestionType;
 }
 
-export interface QuestionnaireResult {
-  metadata: QuestionnaireMetadata;
-  answers: QuestionnaireAnswer[];
+export interface IQuestionnaireResult {
+  metadata: IQuestionnaireMetadata;
+  answers: IQuestionnaireAnswer[];
 }
