+++ src/shared/ui/template/TemplateAnalysis.tsx
@@ -1,18 +1,18 @@
 "use client";
 import React, { useCallback, useEffect, useState } from "react";
-import { Clock, CheckCircle, Filter, Search, Eye, Loader2 } from "lucide-react";
-import { useNavigate } from "react-router-dom";
-import { IAnalysis } from "../../../../core/models/analysis";
+import { Filter, Search, Loader2 } from "lucide-react";
 import { motion } from "motion/react";
+import { IAnalysis } from "../../../core/models/analysis";
+import TableAnalysis from "../components/tables/TableAnalysis";
 
-interface TableSearchAnalysisProps {
+interface TemplateAnalysisProps {
     analysis: IAnalysis[];
     loading: boolean;
     error: string | null;
     role: number | null;
 }
 
-const TableSearchAnalysis: React.FC<TableSearchAnalysisProps> = ({
+const TemplateAnalysis: React.FC<TemplateAnalysisProps> = ({
     loading,
     error,
     analysis,
@@ -21,7 +21,6 @@ const TableSearchAnalysis: React.FC<TableSearchAnalysisProps> = ({
     const [filter, setFilter] = useState<"all" | "pending" | "completed">("all");
     const [searchTerm, setSearchTerm] = useState("");
     const [filteredAnalysis, setFilteredAnalysis] = useState<IAnalysis[]>([]);
-    const navigate = useNavigate();
 
     const filterAnalysis = useCallback(() => {
         let filtered = analysis;
@@ -49,63 +48,6 @@ const TableSearchAnalysis: React.FC<TableSearchAnalysisProps> = ({
     useEffect(() => {
         filterAnalysis();
     }, [filterAnalysis]);
-
-    const getStateBadge = (state: string) => {
-        switch (state?.toUpperCase()) {
-            case "PENDING":
-                return (
-                    <span className="badge badge-warning p-2 badge-sm gap-1 text-xs">
-                        <Clock className="h-3 w-3" />
-                        <span className="hidden sm:inline">Pendiente</span>
-                        <span className="sm:hidden">Pend.</span>
-                    </span>
-                );
-            case "CHECKED":
-            case "COMPLETED":
-                return (
-                    <span className="badge badge-success p-2 badge-sm gap-1 text-xs">
-                        <CheckCircle className="h-3 w-3" />
-                        <span className="hidden sm:inline">Completado</span>
-                        <span className="sm:hidden">Comp.</span>
-                    </span>
-                );
-            default:
-                return (
-                    <span className="badge badge-neutral p-2 badge-sm text-xs">
-                        Desconocido
-                    </span>
-                );
-        }
-    };
-
-    const getColorBadge = (color: string) => {
-        switch (color) {
-            case "verde":
-                return (
-                    <span className="badge badge-success p-2 badge-sm text-xs">Verde</span>
-                );
-            case "amarillo":
-                return (
-                    <span className="badge badge-warning p-2 badge-sm text-xs">Amarillo</span>
-                );
-            case "rojo":
-                return <span className="badge badge-error p-2 badge-sm text-xs">Rojo</span>;
-            default:
-                return (
-                    <span className="badge badge-neutral p-2 badge-sm text-xs">
-                        Sin clasificar
-                    </span>
-                );
-        }
-    };
-
-    const handleViewDetails = (analysis: IAnalysis) => {
-        if (role === 3) {
-            navigate(`/a/analysis/${analysis.analysisId}`);
-        } else if (role === 1) {
-            navigate(`/m/analysis/${analysis.analysisId}`);
-        }
-    };
 
     if (loading)
         return (
@@ -261,106 +203,13 @@ const TableSearchAnalysis: React.FC<TableSearchAnalysisProps> = ({
                 </div>
             </div>
 
-            {/* Tabla principal */}
-            <div className="card bg-base-100 shadow-sm border border-base-200">
-                <div className="card-body p-3 md:p-6">
-                    <h2 className="card-title mb-4 text-lg">Análisis realizados</h2>
-
-                    {filteredAnalysis.length > 0 ? (
-                        <>
-                            {/* Desktop */}
-                            <div className="hidden lg:block overflow-x-auto">
-                                <table className="table table-zebra">
-                                    <thead>
-                                        <tr>
-                                            <th>ID</th>
-                                            <th>Cliente</th>
-                                            <th>Asesor</th>
-                                            <th>Categoría</th>
-                                            <th>Estado</th>
-                                            <th>Riesgo</th>
-                                            <th>Fecha</th>
-                                            <th>Acciones</th>
-                                        </tr>
-                                    </thead>
-                                    <tbody>
-                                        {filteredAnalysis.map((a) => (
-                                            <tr key={a.analysisId}>
-                                                <td>{a.analysisId}</td>
-                                                <td>{a.clientName}</td>
-                                                <td>{a.asesorName || 'Sin Asignar'}</td>
-                                                <td>{a.categoria}</td>
-                                                <td>{getStateBadge(a.status)}</td>
-                                                <td>{getColorBadge(a.colorSemaforo)}</td>
-                                                <td>
-                                                    {new Date(a.timeWhenSolved).toLocaleDateString()}
-                                                </td>
-                                                <td>
-                                                    <button
-                                                        className="btn btn-info btn-xs gap-1"
-                                                        onClick={() => handleViewDetails(a)}
-                                                    >
-                                                        <Eye className="h-3 w-3" />
-                                                        <p className="whitespace-nowrap">Ver</p>
-                                                    </button>
-                                                </td>
-                                            </tr>
-                                        ))}
-                                    </tbody>
-                                </table>
-                            </div>
-
-                            {/* Mobile */}
-                            <div className="lg:hidden space-y-3">
-                                {filteredAnalysis.map((a) => (
-                                    <div key={a.analysisId} className="card bg-base-200 p-4">
-                                        <div className="space-y-2">
-                                            <div className="flex justify-between items-start">
-                                                <div className="flex-1 min-w-0">
-                                                    <span className="text-xs text-base-content/70 whitespace-nowrap">
-                                                        #{a.analysisId} · {new Date(a.timeWhenSolved).toLocaleDateString()}
-                                                    </span>
-                                                    <h3 className="text-md truncate font-bold">
-                                                        {a.clientName}
-                                                    </h3>
-                                                    <h3 className="text-sm capitalize truncate font-normal text-xs">
-                                                        <strong>Categoría: </strong>{a.categoria}
-                                                    </h3>
-                                                    <h3 className="text-sm capitalize truncate font-normal text-xs">
-                                                        <strong>Asesor: </strong>{a.asesorName || "Sin Asignar"}
-                                                    </h3>
-                                                </div>
-                                                <div className="flex-shrink-0">
-                                                    {getStateBadge(a.status)}
-                                                </div>
-                                            </div>
-
-                                            <div className="flex justify-between items-center">
-                                                <div className="flex-shrink-0">
-                                                    {getColorBadge(a.colorSemaforo)}
-                                                </div>
-                                                <button
-                                                    className="btn btn-primary btn-xs gap-1 flex-shrink-0"
-                                                    onClick={() => handleViewDetails(a)}
-                                                >
-                                                    <Eye className="h-3 w-3" />
-                                                    Ver
-                                                </button>
-                                            </div>
-                                        </div>
-                                    </div>
-                                ))}
-                            </div>
-                        </>
-                    ) : (
-                        <div className="text-center py-8 text-base-content/50">
-                            No se encontraron análisis
-                        </div>
-                    )}
-                </div>
-            </div>
+            {/* Tabla análisis */}
+            <TableAnalysis
+                analysis={filteredAnalysis}
+                role={role}
+            />
         </div>
     );
 };
 
-export default TableSearchAnalysis;
+export default TemplateAnalysis;
